Moved files with new algorithm to their propper place.
